added eventController
